run new test images in CI
